[Canary Compatibility] - Modifications for support with canary engine
Added the lua npcs from the canary server and removed the xml npcs.
Changed jiddo lib to work with lua npcs.
Module shop has been removed (a big improvement was made where the shop is now in npcType, more easily and is called directly from source).

Added back storage of npcs spawns in the world folder (there was also a reorganization, where npcs and monsters will be in separate files from now on, as it will be possible to see npcs on the map).

Improvement on register_bestiary_charm and bestiary/charms.lua (formerly modules/bestiary_charms)

Several other modifications and adaptations to make compatibility possible.